# openapi.yml
openapi: "3.0.3"
info:
  title: "MyLittleShopping API"
  version: 0.0.1
paths:
...
  /wishProducts:
    get:
      description: "Get all products per user wishlist"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/wishProduct"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
  /wishProduct:
    post:
      responses:
        201:
          description: "Product added to wishlist"
          content:
            text/plain:
              schema:
                type: string
        400:
          description: "Bad Json Format"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                productId:
                  type: integer
                quantity:
                  type: integer
                availableQuantity:
                  type: integer
              required:
                - productId
                - quantity
                - availableQuantity

    put:
      description: "Modify the quantity of a product in a wishlist"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                items:
                  allOf:
                    - $ref: "#/components/schemas/wishProduct"
        400:
          description: "Bad Json Format"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                sellerUsername:
                  type: string
                productName:
                  type: string
                quantity:
                  type: integer

    delete:
      description: "Allow a buyer to remove a product from his wishlist"
      responses:
        200:
          description: "Product deleted from wishlist"
          content:
            text/plain:
              schema:
                type: string
        400:
          description: "Bad Json Format"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                sellerUsername:
                  type: string
                productName:
                  type: string
              required:
                - sellerUsername
                - productName
...
  /disable:
    put:
      description: "Admin route for user deactivation"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
      responses:
        200:
          description: "Account disabled"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
  /deactivate:
    put:
      description: "Allow a user to deactivate his profile"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
      responses:
        200:
          description: "Account disabled"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
  /syncAccount:
    get:
      description: "Refresh front for caching data"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/user"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
components:
  schemas:
...
    user:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        email:
          type: "string"
        firstname:
          type: "string"
        username:
          type: "string"
        lastname:
          type: "string"
        password:
          type: "string"
        birthdate:
          type: string
          format: date
        role:
          type: "string"
          description: "Role of a user"
          enum: ["buyer", "seller", "admin"]
        activated:
          type: "boolean"
        created_at:
          type: string
          format: date
        updated_at:
          type: string
          format: date
...
    wishProduct:
      properties:
        id:
          type: "integer"
          format: "int64"
        productId:
          type: "integer"
          format: "int64"
        ownerId:
          type: "integer"
          format: "int64"
        quantity:
          type: "integer"
          format: "int64"
        created_at:
          type: string
          format: date
        updated_at:
          type: string
          format: date
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT